backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic-settings>=2.7.0

# PDF Processing
pypdf2>=3.0.1
pdfplumber>=0.11.0

# RAG & Vector Database
# Note: Using SimpleVectorStore (sentence-transformers) instead of ChromaDB for Python 3.14 compatibility
sentence-transformers>=3.3.0
scikit-learn>=1.5.0

# AI/LLM
groq>=0.14.0

# Translation - Google Translate (Simple and Reliable)
deep-translator>=1.11.4

# Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0

# Translation - IndicTrans2 (AI4Bharat for Indian Languages)
# Note: Requires HuggingFace authentication - see INDICTRANS2_SETUP.md
indic-nlp-library>=0.92
sentencepiece>=0.2.0
sacremoses>=0.1.1
transformers>=4.50.0
torch>=2.0.0
huggingface-hub>=0.26.0

# Utilities
python-dotenv>=1.0.1
